Delete some _ in variable and create API for Company

+++ backendBoss/src/main/java/com/boss/backend/controller/UserController.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,7 +27,7 @@ public class UserController {
 
 	@Autowired
 	UserService userService;
-	
+
 	@GetMapping("/hello")
 	ResponseEntity<String> hello() {
 		return new ResponseEntity<>("Hello World!", HttpStatus.BAD_REQUEST);
@@ -40,57 +39,23 @@ public class UserController {
 		return userService.retrieveUser();
 	}
 
-	// Get the user id by user name
+	// Get the user detail by user name
 	@GetMapping("/getUsers/{username}")
 	public ResponseEntity<?> getUserByUsername(@PathVariable(value = "username") String username) {
-		Optional <DAOUser>  user = userService.retrieveUser(username);
+		Optional<DAOUser> user = userService.retrieveUser(username);
 		JSONObject responseJson = new JSONObject();
-		if(!user.isPresent()) {
+		if (!user.isPresent()) {
 			responseJson.put("status", "Not Found this usernames");
 			return new ResponseEntity<JSONObject>(responseJson, HttpStatus.BAD_REQUEST);
 		}
-			 
-		responseJson.put("status", "Found username "+username);
+
+		responseJson.put("Output", user);
 		return new ResponseEntity<JSONObject>(responseJson, HttpStatus.OK);
 	}
 
-//	
-//	@PutMapping("/putUserID/{id}")
-//	public ResponseEntity<?> updateUser(@RequestBody UserDTO user, @PathVariable int id) {
-//		DAOUser usernameIn = userDao.findById(id);
-//		JSONObject responseJson = new JSONObject();
-//
-//		if (usernameIn == null) {
-//			responseJson.put("status", "Not found this ID");
-//			return new ResponseEntity<JSONObject>(responseJson, HttpStatus.BAD_REQUEST);
-//		}
-//		if (user.getUsername() != null)
-//			usernameIn.setUsername(user.getUsername());
-//		if (user.getPassword() != null)
-//			usernameIn.setPassword(bcryptEncoder.encode(user.getPassword()));
-//		if (user.getFirstname() != null)
-//			usernameIn.setFirstname(user.getFirstname());
-//		if (user.getLastname() != null)
-//			usernameIn.setLastname(user.getLastname());
-//		if (user.getAddress() != null)
-//			usernameIn.setAddress(user.getAddress());
-//		if (user.getEmail() != null)
-//			usernameIn.setEmail(user.getEmail());
-//		if (user.getMobile_no() != null)
-//			usernameIn.setMobile_no(user.getMobile_no());
-//		if (user.getCOMPANY_ID() != null)
-//			usernameIn.setCompany_id(user.getCOMPANY_ID());
-//
-//		userDao.save(usernameIn);
-//
-//		responseJson.put("status", "Update complete");
-//		return new ResponseEntity<JSONObject>(responseJson, HttpStatus.OK);
-//	}
-
-	
 	@PutMapping("/putUserID/{id}")
 	public ResponseEntity<?> updateUser(@RequestBody UserDTO user, @PathVariable int id) {
-		DAOUser usernameIn = userService.updateUser(id,user);
+		DAOUser usernameIn = userService.updateUser(id, user);
 		JSONObject responseJson = new JSONObject();
 
 		if (usernameIn == null) {
@@ -102,11 +67,18 @@ public class UserController {
 		return new ResponseEntity<JSONObject>(responseJson, HttpStatus.OK);
 	}
 
-	
-	
 	@DeleteMapping("/deleteUser/{id}")
-	public void deleteUser(@PathVariable int id) {
+	public ResponseEntity<?> deleteUser(@PathVariable int id) {
+		Optional<DAOUser> CheckuserId = userService.findUserId(id);
+		JSONObject responseJson = new JSONObject();
+		if (!CheckuserId.isPresent()) {
+			responseJson.put("status", "Not Found this userId");
+			return new ResponseEntity<JSONObject>(responseJson, HttpStatus.BAD_REQUEST);
+		}
+
 		userDao.deleteById(id);
+		responseJson.put("status", "UserId " + id + " is deleted");
+		return new ResponseEntity<JSONObject>(responseJson, HttpStatus.OK);
 	}
 
 }
